YouTube-Summarizer/backend/requirements-backend.txt
# Core dependencies as per your `pyproject.toml`
accelerate==0.33.0
numpy==1.26.4
websockets==13.1
torch==2.0.0
torchaudio==2.0.1
torchvision==0.15.1
huggingface-hub==0.24.7

# Backend-specific dependencies
fastapi==0.110.1
uvicorn==0.22.0
pyannote-audio==3.1.1
uuid==1.30
yt-dlp==2024.10.7

# Git-based packages
openai-whisper @ git+https://github.com/openai/whisper.git
whisperx @ git+https://github.com/federicotorrielli/BetterWhisperX